lecture7/pyproject.toml
[project]
name = "lecture7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
]
